fix : 아키텍처 구조 모두 수정
