--- lib/data/repository/recipe_repository.dart
@@ -1,6 +1,0 @@
-// lib/domain/repository/recipe_repository.dart
-import 'package:recipe_app/data/model/recipe_model.dart';
-
-abstract class RecipeRepository {
-  Future<List<Recipe>> getRecipes();
-}
