+++ lib/domain/repository/recipe_repository.dart
@@ -1,0 +1,6 @@
+// lib/domain/repository/recipe_repository.dart
+import 'package:recipe_app/domain/entity/recipe_model.dart';
+
+abstract class RecipeRepository {
+  Future<List<Recipe>> getRecipes();
+}
