+++ lib/presentation/widget/custom_text_field.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:recipe_app/app/ui/app_text_styles.dart';
+import 'package:recipe_app/app/ui/color_style.dart';
+
+class CustomTextField extends StatelessWidget {
+  final String hintText;
+
+  const CustomTextField({super.key, required this.hintText});
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      decoration: InputDecoration(
+        hintText: hintText,
+        hintStyle: AppTextStyles.extraSmallRegular(color: ColorStyle.gray4),
+        contentPadding: const EdgeInsets.symmetric(
+          horizontal: 16,
+          vertical: 14,
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+          borderSide: const BorderSide(color: ColorStyle.gray4, width: 1.5),
+        ),
+      ),
+    );
+  }
+}
